+++ NEWS.md
@@ -1,6 +1,8 @@
 Version 0.7.6
  - change: don't auto insert a new line when input an enter between '(' and ')' or between '[' and ']' (indent instead)
  - enhancement: the line containing '}' will use the indents of the matching '{' line, instead of just unindent one level
+ - enhancement: the line containing 'public:' / 'private:' / 'protected:' / 'case *:' will use of indents of the surrounding '{' line, instead of just unindent one level
+ - enhancement: correctly handle auto indents for multi-level embedding complex statements like 'for(...) if (...) printf();
 
 Version 0.7.5
  - enhancement: more accurate auto indent calculation